pyproject.toml
[project]
name = "hems"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "langchain>=1.1.3",
    "langchain-core>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.5",
    "mango-agents>=2.1.5",
    "numpy>=1.24.0",
    "pip>=25.3",
    "pydantic>=2.12.5",
    "pyvis>=0.3.2",
    "uvicorn>=0.38.0",
]
